requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
distro==1.9.0
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
Markdown==3.7
outcome==1.3.0.post0
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.26.1
setuptools==75.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
youtube-transcript-api==0.6.3
